+++ src/ZPublisher/HTTPResponse.py
@@ -97,7 +97,11 @@ _CRLF = re.compile(r'[\r\n]')
 
 
 def _scrubHeader(name, value):
-    if not isinstance(value, str):
+    if isinstance(value, bytes):
+        # handle ``bytes`` values correctly
+        # we assume that the provider knows that HTTP 1.1 stipulates ISO-8859-1
+        value = value.decode('ISO-8859-1')
+    elif not isinstance(value, str):
         value = str(value)
     return ''.join(_CRLF.split(str(name))), ''.join(_CRLF.split(value))